+++ src/l2-proxy-version.h
@@ -1,4 +1,4 @@
 #ifndef L2_PROXY_VERSION_H
 #define L2_PROXY_VERSION_H
-#define VERSION "1.0.4-8081470"
+#define VERSION "1.0.4-ab8d5c1"
 #endif